# Speed up scans of RLE compressed data