This PR reworks the RLE scanning code so that it is more optimized, speeding it up significantly. The exact performance improvement depends on the length of the runs. Below are queries from ClickBench that are bottlenecked by scans speed that benefit from this:

| query |  New  | Main  | Speedup |
|-------|-------|-------|--------:|
| q02   | 0.007 | 0.015 | 2.14    |
| q04   | 0.022 | 0.029 | 1.32    |
| q08   | 0.006 | 0.015 | 2.5     |
| q20   | 0.010 | 0.016 | 1.6     |